+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `SubstanceParameter` encodings are now computed exclusively with the
+  `scikit-fingerprints` package, granting access to all fingerprints available therein
+
+### Deprecations
+- `SubstanceEncoding` value `MORGAN_FP`. Please use `ECFP` with 1024 bits and
+  radius of 4
+- `SubstanceEncoding` value `RDKIT`. Please use `RDKIT2DDESCRIPTORS`.
+
 ## [0.11.1] - 2024-10-01
 ### Added
 - Continuous linear constraints have been consolidated in the new
@@ -16,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Unsafe name-based matching of columns in `get_comp_rep_parameter_indices`
 
-### Deprecated
+### Deprecations
 - `ContinuousLinearEqualityConstraint` and `ContinuousLinearInequalityConstraint`
   replaced by `ContinuousLinearConstraint` with the corresponding `operator` keyword